Squashed commit of the following:

commit 5c8104ebe65146c9315ae4a5a52e8fa3843b19c8
Author: Bhagya Tholpady <bbantwal@cisco.com>
Date:   Tue Mar 17 15:19:57 2020 -0400

    main: check if flow state is blocked while applying verdicts

commit c7f9f5def1d0a42695a0a1e6f9229a443f329b4b
Author: Bhagya Tholpady <bbantwal@cisco.com>
Date:   Tue Dec 10 15:14:37 2019 -0500

    codecs: add new proto bit for udp tunneled traffic

commit 82313413cf6f4152915ffce33fc2f41118bbbd1c
Author: Bhagya Tholpady <bbantwal@cisco.com>
Date:   Thu Aug 30 13:41:51 2018 -0400

    codecs: add vxlan codec

    move gtp_ports and deep_teredo_inspection from
    snort config to UdpCodecConfig

+++ src/packet_io/sfdaq.cc
@@ -371,7 +371,7 @@ bool SFDAQ::can_replace()
     return local_instance && local_instance->can_replace();
 }
 
-bool SFDAQ::get_tunnel_bypass(uint8_t proto)
+bool SFDAQ::get_tunnel_bypass(uint16_t proto)
 {
     return local_instance && local_instance->get_tunnel_bypass(proto);
 }
